Set update_field for workflow state of leave application
